backend/requirements.txt
fastapi
uvicorn
eodag
rasterio
rio-cogeo
numpy
pandas
geopandas
shapely
pyproj
SQLAlchemy
geoalchemy2
alembic
minio
prefect>=2.0
pydantic
python-multipart
psycopg2-binary
boto3
typer
ruff
black
isort
mypy
pytest
pytest-asyncio
httpx
Pillow
